refactor(mf): start-menu - scss into css
