--- micro-frontends/src/mf-start-menu/src/ui/modules/button-start/button-start.module.scss
@@ -1,7 +1,0 @@
-.button-start {
-    // TODO: This code is temporary and should be removed after migrattion to panda CSS
-    font: 11px/11px MSSansSerifBold, sans-serif !important;
-    column-gap: 4px !important;
-    min-width: 55px !important;
-    padding: 0 6px !important;
-}
