+++ micro-frontends/src/mf-start-menu/src/ui/modules/menu-item/content/content.module.css
@@ -1,0 +1,5 @@
+.content {
+	display: flex;
+	align-items: center;
+	column-gap: var(--ds-spacing-sm);
+}
