--- micro-frontends/src/mf-start-menu/src/ui/modules/menu-item/content/content.module.scss
@@ -1,5 +1,0 @@
-.content {
-    display: flex;
-    align-items: center;
-    column-gap: 5px;
-}
