+++ micro-frontends/src/mf-start-menu/src/ui/modules/menu-item/menu-item.module.css
@@ -1,0 +1,3 @@
+.menu-item {
+	width: 100%;
+}
