--- micro-frontends/src/mf-start-menu/src/ui/modules/menu-item/menu-item.module.scss
@@ -1,3 +1,0 @@
-.menu-item {
-    width: 100%;
-}
