--- micro-frontends/src/mf-start-menu/src/ui/modules/menu-item/wrapper/wrapper.module.scss
@@ -1,33 +1,0 @@
-.menu-item-wrapper {
-  position: relative;
-  display: flex;
-  align-items: center;
-  cursor: pointer;
-  padding: 5px;
-  min-width: 135px;
-
-  &:hover {
-    @include hover {
-      background-color: $cl-blue33;
-
-      & > div > p {
-        color: var(--text-color-2);
-      }
-    }
-  }
-}
-
-.type-group {
-  position: relative;
-
-  &::after {
-      content: '';
-      position: absolute;
-      top: 50%;
-      right: 5px;
-      transform: translateY(-50%);
-      border-top: 4px solid transparent;
-      border-bottom: 4px solid transparent;
-      border-left: 4px solid $cl-black;
-    }
-}
