+++ micro-frontends/src/mf-start-menu/src/ui/modules/menu-items/wrapper/wrapper.tsx
@@ -1,5 +1,6 @@
+import { getClassNames } from "@windows98/toolkit"
 import type { FunctionComponent } from "react"
-import styles from "./wrapper.module.scss"
+import styles from "./wrapper.module.css"
 import type { MenuItemsWrapperProps } from "./wrapper.type"
 
 export const MenuItemsWrapper: FunctionComponent<MenuItemsWrapperProps> = ({
@@ -8,7 +9,10 @@ export const MenuItemsWrapper: FunctionComponent<MenuItemsWrapperProps> = ({
 }) => {
 	return (
 		<div
-			className={`${styles["menu-items-wrapper"]} ${styles[`variant-${variant}`]}`}
+			className={getClassNames([
+				styles["menu-items-wrapper"],
+				styles[`variant-${variant}`],
+			])}
 			data-testid={`mf-start-menu-menu-items-wrapper-${variant}`}
 		>
 			{children}
